arithmetic: Strip tags on all arithmetic operations
Strip tags in all arithmetic operations, including on arrays/lists.

This introduces a small incompatibility with HP. On HP calculators,
`:A:{1 2} :B:{3 4}` gives `{ 1 2 3 4 }`, but
`{1 2} :B:{3 4}` gives `{ 1 2 :B:{ 3 4 }}`.

This is inconsistent, complicated and causes operations on tagged
arrays to misbehave. Stripping tags in all cases makes things much
more consistent and simpler. This means that adding a tagged object to
a list requires the `→List` operation.

Fixes: #1358

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>