constants: Evaluate constant body in `→Num`
In order to be able to accomodate computed constants, as per
CODATA2022, where some physical constants are defined not numerically
but through a computation from other constants, we need to be able to
evaluate the body of expressions in the `→Num` command.

In the process, added evaluation for `xlib` objects, and simplified
the code to evaluate expressions.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Suggested-by: Wiljea <143950411+Wiljea@users.noreply.github.com>