cleaner: Clear the error command if necessary
When we purge temporaries in the `cleaner` class after a lengthy
operation, it is possible that the runtime command is in the area of
temporaries that we are cleaning up. In that case, simply clear the
error command. We are at the end of the commadn execution, and we
should normally no longer need the command name in that case. What was
saved was most likely a command in an expression.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>