runtime: Fix crash for deep recursion with for loops
DB48x maintains a call stack when evaluating objects.

To avoid moving root pointers too often during object evaluation, it
grows and shrinks the call stack in 32 pointers increments, which
roughly correspond to 16 calls or 8 conditional loops.

The order between popping the `Returns` pointer and dropping 32
pointers from the call stack matters. If it is not respected, this can
lead to stack corruption because for example the `Locals` pointer will
move above the `Returns` pointer undetected.

Three independent problems existed in the update order of this call
stack, one of which was triggered during the evaluation of the test
program given by user @skai111.

The original call stack adjustment technique was predicated on modulo
values, which is fragile and could trip at the wrong offset while
failing to detect a "missed" step. The logic was replaced with a more
robust one simply comparing `Returns` and `CallStack`, which is the
correct (and robust) test.

Additionally, on simulators, enabled a `runtime_invariants` class that
checks runtime invariants on entry and exit of key runtime functions.

Fixes: #1362

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>