simulator: Make the gcptr thread safe
Now that the simulator uses a `for` loop on a list to detect which key
is being pressed, this triggers a relatively frequent corruption of
the GCPtr list. One possible manifestation was an infinite loop on the
list of safe pointers.

Add a simulator-only locking mechanism. On the calculator, we don't
want to pay the overhead of even lock-free list management.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>