keyboard: Disable `TIMER0` to avoid spurious "long presses"
When we wait without a key being pressed, `TIMER0` expires.
It does not mean a repeat if there is no key pressed.

After the rework of the power management, reducing the frequency of
animations controled by `TIMER1`, this could lead to a scenario where
`TIMER0` would expire first without any key being pressed, and would
then remain active and expired until the next key press. The next key
press would the incorrectly be interpreted as a long press.

This could also lead to scenarios where there was no key, but `TIMER0`
would repeatedly kick us out until a key was pressed and it was reset,
because nothing was clearing it. This could explain the battery drain
that some people have reported.

The solution is to check if there is a key, and if no key is pressed,
to stop `TIMER0` and ignore any timeouts initiated when the previous
key had been pressed.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>