display: Add support for hardware-accelerated display refresh
Use the DMCP built-in functions for DMA-enabled display refresh.
This is configurable using the `SoftwareDisplayRefresh` to revert to
software, or `DMCPDisplayRefresh` to enable background refresh.

Added measurements of the time spent drawing (with a separate
measurement for the stack), and of the time spent refreshing.
With DMA-enabled screen refresh, the time spent in refresh is
practically zero. This also appears to reduce battery drain quit
significantly, as evidenced by the power voltage shown when using a
fresh battery (it was showing 2.7V before, now 2.9V).

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>