tests: Avoid concurrent pixmap updates on  main thread
The `-Texamples` test sometimes fail with a crash complaining about a
double free in `QPixmap`. This happens for programs that do screen
updates and then call `Wait`, causing another refresh. The
`postToThread` ensures we defer it to another thraed, but they do not
prevent the RPL thread from attempting to re-update the pixmap;

The solution is simply to wait until the main thread has completed the
update (and grabbed a copy of the QPixmap).

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>