files: Automate tracking of open files
DMCP has a semi-absurd limitation where you can only have one open
file at a time.

To address this, we used to have a `file_closer` class to
automatically close an existing file in situations where two files
might be opened at the same time. However, this was error prone and
hard to maintain.

By simply maintaining a linked list of open file, we can actually open
and close files automatically and not bother with the `file_closer`
class anymore.

While this code was being reworked, replaced the boolean flag for
reading or writing with an enumeration.

Files: #1383

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>