random: Fix bias in integer version of `Random`
The random number generator for integer ranges was giving a biased
result because it was using an incorrect extent. For example, for a
range of `2` to `4`, the extent would be 2 and an offset of 0.5 was
applied. This would cause values `2` and `4` to happen less often than
`3`.

Since in that case, we want to generate 3 values, use an extent of 3
instead, and no offset, but use the `floor` function to ensure we
round to the lower value.

Fixes: #1378