hwfp: Add explicit `F` or `D` suffix
This commit introduces an explicit syntax with a `D` or `F` suffix for
binary IEEE754 floating-point values in 64-bit and 32-bit respectively.
In other words, `1.5` is always a (variable-precision) decimal value,
`1.5F` is always a 32-bit IEEE754 binary floating-point value, and
`1.5D` is always a 64-bit IEEE754 binary floating-point value. `F`
stands for `float` and `D` for `double`, the name of these types in
most modern C implementations.

The issue described in #1394 highlighted that the way a program was
parsed before depended on the `HardwareFloatingPoint` setting. For
example, `« 1.2 + »`, when entered with hardware floating-point
enabled and `Precision` set to 7, would end up with a program that
contains the `1.20000 00476 8` constant instead of the expected
`1.2`. This was obviously bad for portability. Conversely, when
`HardwareFloatingPoint` was disabled, it would not be possible to
enter a binary floating-point constant.

In addition to eliminating these bugs, the new syntax makes it clear
what type of value is on the stack, and makes it possible to enter
mixed expressions such as `'1.2F+4.5D-4.2'`.

Fixes: #1394

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>