algebraic: `decimal_promotion` should not fail silently
The code as written resulted in cases where the input argument to the
`decimal_romotion` function would be nullified when the conversion
fails. This happens for example in case of infinities that are not
correctly translated to decimal.

However, its use in the arithmetic code implies that the function
should not nullify a non-null input even if it returns `false`.
Modify the internal logica accordingly. This "unsilences" the internal
failure that happens when running `'COMB(9999., 555.)'` with
`HardwareFloatingPoint` enabled and `NumericalResults` set.

Fixes: #1397

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>