ui: Implement `Prompt` and `Run` commands
The `Prompt` command behaves like on the HP48SX or HP50G.
It halts the program, which you can resume with `Continue` like on
the HP50G. However, because `Continue` is not available through a
direct key assignment, the DM42 R/S key is changed from `Evaluate` to
`Run`.

The `Run` command is a DB48x-specific addition, intended to make
`Prompt` easier to use. It behaves like `Evaluate`, unless a program
was interrupted, in which case it resumes it (behaving like
`Continue`).

You can use `Prompt` in a program, and once you are done with
the operations that you are `Prompt`-ed to perform, you can simply hit
the "space" key to resume execution.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>