integer: Do not read character beyond input
Characters could be read beyond the input source code during integer
parsing. In addition, the pointer to the current character was
restored after a potential GC using a GC-safe pointer, which would
grow if it was pointing to the last byte in the temporary area.

In both cases, the effect would be to compute an incorrect length for
the parsed integer if it was at the end of the input line. This was
detected by the `Input` validation tests.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>