convert: Evaluate arguments if necessary
Consider the following code:

```
α=0.00729735256434 Urα=1.6E-10
μ0='CONVERT(4*Ⓒπ*α*Ⓒℏ/(Ⓒqe^2*Ⓒc);1_H/m)'
```

The conversion requires the evaluation of the expression to get a unit
object. Otherwise, `CONVERT` fails with a `Bad argument type`.

This commit implements the evaluation for all unit values. Presumably,
one could also do something like `'CONVERT(A;B)'` where `A` and `B`
are variables or programs.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>