equations: Change initial guess for Plate Capacitor
The initial guess for `d` was so large that the algorithm fails to
converge with the default number of iterations allowed.

The situation is similar to solving `1-1/x` with an initial value of
`1E9`: the area of the function is so flat that far out that following
the slope leads you to a large negative value, making things worse. It
then takes a while to reconverge. Ultimately, we evaluate the function
more than `SolverIterations` times, so the solver emits `No solution`.

The HP50G and HP48SX behave even worse with similar input: they just
return a value without any error message. For example,
`'1-1/X' 'X' 1E9 ROOT` returns `11457.33451` on both machines, which
is ridiculously wrong (the value is `0.99991 27196 64` and not `0`).

Fixes: #1399

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>