arithmentic: Evaluate unit expressions in add/sub arguments
If the argument to `+` or `-` is an expression that generates a unit
value, we need to evaluate it to check if conversion is possible.

For example: `'1_m/s-(1_A)÷((8.5⁳28_(m↑3)⁻¹)·Ⓒqe·Ⓒπ·(0.01_cm↑2))'`
has a right-hand side that is *not* a unit value, so we were not
correctly evaluating it, resulting in a bogus `Inconsistent units`
message.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>