program: When evaluating in synchronous mode, evaluate names
When we evaluate a name like `A` in `sync` mode, we would run
`symbol::evaluate()` which, if the content of `A` was itself a
program (including an equation), would call `run_program`. This pushes
the program to evaluate, but that program would then not be evaluated
via `run_loop`. This run loop only happens if the argument to
`program::run` was itself a program.

The result is that in the case of algebraic evaluation, the evaluation
would fail with an `Invalid algebraic` message because the stack was
not updated. The case was detected in solver tests, see #1400, after
solving a few other problems with that case.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>