equations: Remove `=0` from equations that have a unit
One equation was written

```
'(ΔP_Pa)/(ρ_(kg/m^3))+((v2_(m/s))^2-(v1_(m/s))^2)/2+Ⓒg*(Δy_m)=0'
```

While it's tolerated in physics to write `=0` like this, from a unit
point of view this is not correct, since the left-hand-side is not
dimensionless.

The solution is to remove the `=0` at the end.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>