tests: Purge the directory between example tests by default
Unless there is a `@ Keep` or `@ Save` in the command, purge the tests
directory so that we get the exact same results as if we were doing
the test first.

This makes the test suite more robus, e.g. we can reorder tests more
easily. This also ensures that if you copy/paste the example from an
empty state, you get the expected result. In some cases, there may be
minute differences because the saved value of some variables causes
the initial search to diverge a little from "default".

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>