equations: Fix capitalization for Angular Mechanics
There is a typo in one of the input variables, `Θ=360_°` (uppercase)
instead of `Θ=360_°` (lowercase).

Before the change in the test harness that cleans up between runs,
this would simply manifest as a wrong value because the value of the
angle was taken from the wrong variable.

After the change in the test harness, this manifested as an error that
it could not solve for all variables. This made it easier to figure
out that variable `Θ` was missing.

It would be nice to improve the detection of the variable for which we
fail to find a solution, but it can be deduced from the variables that
have been written.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>