solver: Improve heuristic when slope is small
Consider the case where you solve `'tan(x)=200'` with initial guess at
1 degree. The solution is `atan(200)`, which is roughly `89.71`.
When we start around 1, the slope of ``tan` is quite small, so that
the default linear interpolation sends us way into the woods,
e.g. around `x=12000`. After that, all bets are off because we are
beyond the current period, and it's only luck if we ever find a
solution (the solver rarely does).

The proposed heuristic is to more carefully move towards the smallest
value in that case. If the function is reasonably uniform, this will
bring us towards a smaller value.  Either we reach a point where the
slope becomes higher and we can then use the faster convergence
algorithm, or we cross positive and negative and we can then bisect.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>