nfunctions: Evaluate symbolic arguments symbolically
When we evaluate something like `'∫(0;π;exp x;x)'` and `x` has a
value, we could end up with the inner expression incorrectly
evaluated, i.e. we would get `'∫(0;π;20.08553 69232;3)'` and then we
would get a `Not an equation or program` error evaluating it.

When evaluating symbolically, wrap symbolic arguments in the
expression so that we get them back during the next evaluation.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>