equations: Remove additional `IC` variable
The test command for the `Bipolar Transistors` was asking to solve for
`IC`, which is an input to the equation and cannot be solved from
within.

In the end, the results for `IE`, `IB` and `VCEsat` are different from
what the HP50G Advanced Reference Manual gives, but DB48x appears to
be correct in its computation. The exponential amplifies the
difference in the ratio `q/k`, and the constants are slightly
different.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>