doc: Document difference parsing `'N'`
On HP calculators, `'N'` is a symbol, not an expression, with the
exception of programs.

In other words, `« 'N' N »` parses as you would expect, like on DB48x,
but `{ 'N' N }` parses as `{ N N }`, which makes it a bit difficult to
programmatically build programs without parsing text.

The DB48x approach appears more consistent and easier, so document it
instead of trying to mimic the exact behaviour of HP calculators.

Fixes: #1412

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>