constants: Add code to manage uncertainties
The standard and relative uncertainties are now represented by two new
type of object. The standard uncertainty of constant `Ⓒg` is `Ⓢg`,
and its  relative uncertainty is `Ⓡg`.

The `StandardUncertainty` command returns the standard uncertainty for
a given constant, which can be identified by a symbol or a text, like
for the `Const` command. The `RelativeUncertainty` returns the
relative uncertainty.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>