integrate: Implement Kahan's improvements to Romberg's method
In his original [HP Journal article][hp] describing the integration
method used in the HP34C, Dr. Kahan indicates (on page 26) that the
algorithm uses a change of variables to avoid uniformly spaced
sampling (reducing resonance or aliasing), and making it possible to
avoid samples at end of the interval.

This technique was preserved in all subsequence HP calculators. It is
explicitly mentioned in the HP15C advanced functions handbook (on page
41). This suggests that this improvement was quite good.

A [user remarked][bc] that this improvement is necessary to integrate
`'sin(x)/x'` over 0...π without getting a divide by zero in zero. The
improved algorithm is able to avoid sampling the endpoints (though it
samples the midpoint, so it does not help if the range is -π...π.

This change improves the results in all test cases but one.

[hp]: https://hparchive.com/Journals/HPJ-1980-08.pdf
[bc]: https://www.hpmuseum.org/forum/thread-20157-post-198839.html#pid198839

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>